# frontend/app/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "TG Bot Admin UI",
    "version": "0.1.0"
  },
  "paths": {
...
    "/cities": {
      "get": {
        "tags": [
          "cities"
        ],
        "summary": "Cities List",
        "operationId": "cities_list_cities_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "text/html": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBasic": []
          }
        ]
      }
    },
    "/cities/new": {
      "get": {
        "tags": [
          "cities"
        ],
        "summary": "Cities New",
        "operationId": "cities_new_cities_new_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "text/html": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBasic": []
          }
        ]
      }
    },
    "/cities/{city_id}/edit": {
      "get": {
        "tags": [
          "cities"
        ],
        "summary": "Cities Edit Page",
        "operationId": "cities_edit_page_cities__city_id__edit_get",
        "security": [
          {
            "HTTPBasic": []
          }
        ],
        "parameters": [
          {
            "name": "city_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "City Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "text/html": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "cities"
        ],
        "summary": "Cities Edit Submit",
        "operationId": "cities_edit_submit_cities__city_id__edit_post",
        "security": [
          {
            "HTTPBasic": []
          }
        ],
        "parameters": [
          {
            "name": "city_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "City Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "text/html": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/cities/create": {
      "post": {
        "tags": [
          "cities"
        ],
        "summary": "Cities Create",
        "operationId": "cities_create_cities_create_post",
        "requestBody": {
          "content": {
            "application/x-www-form-urlencoded": {
              "schema": {
                "$ref": "#/components/schemas/Body_cities_create_cities_create_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBasic": []
          }
        ]
      }
    },
    "/cities/{city_id}/settings": {
      "post": {
        "tags": [
          "cities"
        ],
        "summary": "Update City Settings",
        "operationId": "update_city_settings_cities__city_id__settings_post",
        "security": [
          {
            "HTTPBasic": []
          }
        ],
        "parameters": [
          {
            "name": "city_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "City Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/cities/{city_id}/owner": {
      "post": {
        "tags": [
          "cities"
        ],
        "summary": "Update City Owner",
        "operationId": "update_city_owner_cities__city_id__owner_post",
        "security": [
          {
            "HTTPBasic": []
          }
        ],
        "parameters": [
          {
            "name": "city_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "City Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/cities/{city_id}/status": {
      "post": {
        "tags": [
          "cities"
        ],
        "summary": "Update City Status Api",
        "operationId": "update_city_status_api_cities__city_id__status_post",
        "security": [
          {
            "HTTPBasic": []
          }
        ],
        "parameters": [
          {
            "name": "city_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "City Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/cities/{city_id}/delete": {
      "post": {
        "tags": [
          "cities"
        ],
        "summary": "Cities Delete",
        "operationId": "cities_delete_cities__city_id__delete_post",
        "security": [
          {
            "HTTPBasic": []
          }
        ],
        "parameters": [
          {
            "name": "city_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "City Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/templates": {
      "get": {
        "tags": [
          "templates"
        ],
        "summary": "Templates List",
        "operationId": "templates_list_templates_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "text/html": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBasic": []
          }
        ]
      }
    },
    "/templates/new": {
      "get": {
        "tags": [
          "templates"
        ],
        "summary": "Templates New",
        "operationId": "templates_new_templates_new_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "text/html": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBasic": []
          }
        ]
      }
    },
    "/templates/create": {
      "post": {
        "tags": [
          "templates"
        ],
        "summary": "Templates Create",
        "operationId": "templates_create_templates_create_post",
        "requestBody": {
          "content": {
            "application/x-www-form-urlencoded": {
              "schema": {
                "$ref": "#/components/schemas/Body_templates_create_templates_create_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBasic": []
          }
        ]
      }
    },
...
    "/templates/{tmpl_id}/edit": {
      "get": {
        "tags": [
          "templates"
        ],
        "summary": "Templates Edit",
        "operationId": "templates_edit_templates__tmpl_id__edit_get",
        "security": [
          {
            "HTTPBasic": []
          }
        ],
        "parameters": [
          {
            "name": "tmpl_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Tmpl Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "text/html": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/templates/{tmpl_id}/delete": {
      "post": {
        "tags": [
          "templates"
        ],
        "summary": "Templates Delete",
        "operationId": "templates_delete_templates__tmpl_id__delete_post",
        "security": [
          {
            "HTTPBasic": []
          }
        ],
        "parameters": [
          {
            "name": "tmpl_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Tmpl Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/cities": {
      "get": {
        "tags": [
          "api"
        ],
        "summary": "Api Cities",
        "operationId": "api_cities_api_cities_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "HTTPBasic": []
          }
        ]
      }
    },
...
    "/api/templates": {
      "get": {
        "tags": [
          "api"
        ],
        "summary": "Api Templates",
        "operationId": "api_templates_api_templates_get",
        "security": [
          {
            "HTTPBasic": []
          }
        ],
        "parameters": [
          {
            "name": "city_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "City Id"
            }
          },
          {
            "name": "key",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Body_cities_create_cities_create_post": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "tz": {
            "type": "string",
            "title": "Tz",
            "default": "Europe/Moscow"
          }
        },
        "type": "object",
        "required": [
          "name"
        ],
        "title": "Body_cities_create_cities_create_post"
      },
...
      "Body_templates_create_templates_create_post": {
        "properties": {
          "text": {
            "type": "string",
            "title": "Text"
          },
          "city_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "City Id"
          },
          "is_global": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Is Global"
          }
        },
        "type": "object",
        "required": [
          "text"
        ],
        "title": "Body_templates_create_templates_create_post"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBasic": {
        "type": "http",
        "scheme": "basic"
      }
    }
  }
}